+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 
 - Add Karpenter deployment manifests
+- provider: drop clusterName (unused attribute)
 
 0.0.1
 ------